+++ .github/workflows/black-linter.yml
@@ -4,13 +4,11 @@ on:
   push:
     paths:
       - "**.py"
-      - "src/backend/**"
     branches:
       - develop
   pull_request:
     paths:
       - "**.py"
-      - "src/backend/**"
     branches:
       - develop
