--- src/backend/.env.example
@@ -1,11 +1,0 @@
-# SECURITY WARNING: don't run with the debug turned on in production!
-DEBUG=True
-
-# SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY=secret
-
-# A list of all the people who get code error notifications.
-ADMINS=John Doe <john@example.com>, Mary <mary@example.com>
-
-# A list of hosts to be allowed for the project
-ALLOWED_HOSTS=127.0.0.1,localhost
