--- src/backend/main/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class MainConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "main"
-
-    def ready(self):
-        import main.signals  # noqa
