--- src/backend/main/fixtures/initial.json
@@ -1,107 +1,0 @@
-[
-    {
-        "model": "main.portionitem",
-        "pk": 1,
-        "fields": {
-            "name": "Carbohydrates",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 2,
-        "fields": {
-            "name": "Fruits & Vegetables",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 3,
-        "fields": {
-            "name": "Protein",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 4,
-        "fields": {
-            "name": "Dairy",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 5,
-        "fields": {
-            "name": "Oils & Fats",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.user",
-        "pk": 1,
-        "fields": {
-            "email": "ben.tullett@portionmate.co.uk",
-            "first_name": "Ben",
-            "last_name": "Tullett",
-            "picture": "",
-            "age": 30,
-            "height": 175,
-            "weight": 70
-        }
-    },
-    {
-        "model": "main.user",
-        "pk": 2,
-        "fields": {
-            "email": "michelle.braun@portionmate.co.uk",
-            "first_name": "Michelle",
-            "last_name": "Braun",
-            "picture": "",
-            "age": 22,
-            "height": 168,
-            "weight": 52
-        }
-    },
-    {
-        "model": "main.user",
-        "pk": 3,
-        "fields": {
-            "email": "brigette.rogers@portionmate.co.uk",
-            "first_name": "Brigette",
-            "last_name": "Rogers",
-            "picture": "",
-            "age": 18,
-            "height": 172,
-            "weight": 54
-        }
-    },
-    {
-        "model": "main.user",
-        "pk": null,
-        "fields": {
-            "email": "john@example.com",
-            "first_name": "John",
-            "last_name": "Doe",
-            "picture": "",
-            "age": null,
-            "height": null,
-            "weight": null
-        }
-    },
-    {
-        "model": "main.user",
-        "pk": null,
-        "fields": {
-            "email": "jane@example.com",
-            "first_name": "Jane",
-            "last_name": "Smith",
-            "picture": "",
-            "age": null,
-            "height": null,
-            "weight": null
-        }
-    }
-]
