--- src/backend/main/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import redirect
-
-
-def index(request):
-    return redirect("/admin")
