--- src/backend/portion_mate/asgi.py
@@ -1,7 +1,0 @@
-import os
-
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "portion_mate.settings")
-
-application = get_asgi_application()
