--- src/backend/portion_mate/wsgi.py
@@ -1,7 +1,0 @@
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "portion_mate.settings")
-
-application = get_wsgi_application()
