+++ src/server/main/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import redirect
+
+
+def index(request):
+    return redirect("/admin")
