+++ src/server/templates/admin/base_site.html
