+++ src/server/templates/rest_framework/base.html
